node-data: move ledger::block in its own module
